feat: use gate owner

+++ app/Repositories/Eloquent/ReplySupportRepository.php
@@ -6,6 +6,7 @@ use App\DTO\Replies\CreateReplyDTO;
 use App\Models\ReplySupport as Model;
 use App\Repositories\Contracts\ReplyRepositoryInterface;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use stdClass;
 
 class ReplySupportRepository implements ReplyRepositoryInterface
@@ -26,13 +27,12 @@ class ReplySupportRepository implements ReplyRepositoryInterface
 
     public function createNew(CreateReplyDTO $dto): stdClass
     {
-        $reply = $this->model->with('user')->create([
+        $reply = $this->model->create([
             'content' => $dto->content,
             'support_id' => $dto->supportId,
             // 'user_id' => auth()::user()->id,
             'user_id' => Auth::user()->id,
         ]);
-        $reply = $reply->with('user')->first();
 
         return (object) $reply->toArray();
     }
@@ -41,6 +41,10 @@ class ReplySupportRepository implements ReplyRepositoryInterface
     {
         if (!$reply = $this->model->find($id)) {
             return false;
+        }
+
+        if (Gate::denies('owner', $reply->user->id)) {
+            abort(403, 'Not Authorized');
         }
 
         return (bool) $reply->delete();
